# packages/schema/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LIGER CLUB API",
    "version": "1.0.0",
    "description": "API for LIGER CLUB MVP"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/proposal": {
      "post": {
        "summary": "Create a new proposal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proposal created"
          }
        }
      }
    },
    "/proposal/{id}": {
      "get": {
        "summary": "Get proposal details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proposal brief"
          }
        }
      }
    },
    "/proposal/{id}/approve": {
      "post": {
        "summary": "Approve a proposal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proposal approved"
          }
        }
      }
    },
    "/proposal/{id}/reject": {
      "post": {
        "summary": "Reject a proposal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proposal rejected"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProposalRequest": {
        "type": "object",
        "properties": {
          "studentId": {"type": "integer"},
          "vertical": {"type": "string"},
          "instrument": {"type": "string"},
          "amount": {"type": "number"}
        },
        "required": ["studentId", "vertical", "instrument", "amount"]
      },
...
    }
  }
}